requirements.txt
numpy>=1.22
pandas>=0.21.0
Pillow==9.5.0
prediction==0.0.3
streamlit==1.20.0
tensorflow>=1.5.0
scikit-learn